fix: add header-stats command
