+++ applications/tari_base_node/src/commands/command_handler.rs
@@ -236,113 +236,6 @@ impl CommandHandler {
         Ok(())
     }
 
-    pub async fn save_header_stats(
-        &self,
-        start_height: u64,
-        end_height: u64,
-        filename: String,
-        pow_algo: Option<PowAlgorithm>,
-    ) -> Result<(), Error> {
-        let mut output = File::create(&filename).await?;
-
-        println!(
-            "Loading header from height {} to {} and dumping to file [working-dir]/{}.{}",
-            start_height,
-            end_height,
-            filename,
-            pow_algo.map(|a| format!(" PoW algo = {}", a)).unwrap_or_default()
-        );
-
-        let start_height = cmp::max(start_height, 1);
-        let mut prev_header = self.blockchain_db.fetch_chain_header(start_height - 1).await?;
-
-        let mut buff = Vec::new();
-        writeln!(
-            buff,
-            "Height,Achieved,TargetDifficulty,CalculatedDifficulty,SolveTime,NormalizedSolveTime,Algo,Timestamp,\
-             Window,Acc.Monero,Acc.Sha3"
-        )?;
-        output.write_all(&buff).await?;
-
-        for height in start_height..=end_height {
-            let header = self.blockchain_db.fetch_chain_header(height).await?;
-
-            // Optionally, filter out pow algos
-            if pow_algo.map(|algo| header.header().pow_algo() != algo).unwrap_or(false) {
-                continue;
-            }
-
-            let target_diff = self
-                .blockchain_db
-                .fetch_target_difficulties_for_next_block(prev_header.hash().clone())
-                .await?;
-            let pow_algo = header.header().pow_algo();
-
-            let min = self
-                .consensus_rules
-                .consensus_constants(height)
-                .min_pow_difficulty(pow_algo);
-            let max = self
-                .consensus_rules
-                .consensus_constants(height)
-                .max_pow_difficulty(pow_algo);
-
-            let calculated_target_difficulty = target_diff.get(pow_algo).calculate(min, max);
-            let existing_target_difficulty = header.accumulated_data().target_difficulty;
-            let achieved = header.accumulated_data().achieved_difficulty;
-            let solve_time = header.header().timestamp.as_u64() as i64 - prev_header.header().timestamp.as_u64() as i64;
-            let normalized_solve_time = cmp::min(
-                cmp::max(solve_time, 1) as u64,
-                self.consensus_rules
-                    .consensus_constants(height)
-                    .get_difficulty_max_block_interval(pow_algo),
-            );
-            let acc_sha3 = header.accumulated_data().accumulated_sha_difficulty;
-            let acc_monero = header.accumulated_data().accumulated_monero_difficulty;
-
-            buff.clear();
-            writeln!(
-                buff,
-                "{},{},{},{},{},{},{},{},{},{},{}",
-                height,
-                achieved.as_u64(),
-                existing_target_difficulty.as_u64(),
-                calculated_target_difficulty.as_u64(),
-                solve_time,
-                normalized_solve_time,
-                pow_algo,
-                chrono::DateTime::from(header.header().timestamp),
-                target_diff.get(pow_algo).len(),
-                acc_monero.as_u64(),
-                acc_sha3.as_u64(),
-            )?;
-            output.write_all(&buff).await?;
-
-            if header.header().hash() != header.accumulated_data().hash {
-                eprintln!(
-                    "Difference in hash at {}! header = {} and accum hash = {}",
-                    height,
-                    header.header().hash().to_hex(),
-                    header.accumulated_data().hash.to_hex()
-                );
-            }
-
-            if existing_target_difficulty != calculated_target_difficulty {
-                eprintln!(
-                    "Difference at {}! existing = {} and calculated = {}",
-                    height, existing_target_difficulty, calculated_target_difficulty
-                );
-            }
-
-            print!("{}", height);
-            io::stdout().flush()?;
-            print!("\x1B[{}D\x1B[K", (height + 1).to_string().chars().count());
-            prev_header = header;
-        }
-        println!("Complete");
-        Ok(())
-    }
-
     pub(crate) fn get_software_updater(&self) -> SoftwareUpdaterHandle {
         self.software_updater.clone()
     }
